extract and document handling of zones
Pull all the zone stuff into its own file.
Document in detail how it works,
and why it looks the way it does.

Add future build breakage to force us to recertify
that the package still works.
