+++ zone_future.go
@@ -1,0 +1,17 @@
+// Copyright 2020 The Inet.Af AUTHORS. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// +build go1.19
+
+package netaddr
+
+It is December, 2020. The Go 1.16 craze is about to sweep the world.
+
+We are not sure this package will continue to function in the future,
+due to the unsafe roguery contained in zone.go.
+
+If you are seeing a compiler error from this file,
+from this not-a-comment comment,
+it is because we have not checked that this package is still safe to use.
+
